requirements.txt
mysql-connector-python
python-dotenv
pdfplumber
pypdf
google-genai
bcrypt
streamlit
pandas
